refactor: extract common filter harness code
